Enable OCSP stapling when ACME_READY = true
Let's Encrypt recommends enabling OCSP stapling in their integration
guide (https://letsencrypt.org/docs/integration-guide/), and we do so
here. The main benefit of OCSP is to avoid an OCSP revokation check
being made when a browser attempts to connect to a site using one of our
ACME certificates.

In the longer term, we might want to:

- Expose this for non-ACME hosts.
- Enable `ssl_stapling_verify`.

The latter in particular should be done with care. While adding an OCSP
staple to our responses (even if it's merely a fail) shouldn't possibly
do harm, it's probably best to leave it up to the client browser to
decide whether the OCSP response is a fail or a pass.