Expose healthcheck endpoint
This exposes a third port on the instance (9000), which can be used for
health checks. This will accept regular HTTP requests and forward them
to a `/healthcheck` endpoint in the underlying app.

If the app responds (regardless of the status), the health check is
considered successful and we return a 200 (which an ELB will interpret
as a pass), and otherwise we return a 502 (which it will interpret as
failure).

All request and response components are discarded (the server never sees
the client's intput, and the client never sees the server response) so
that we don't send anything over the wire unencrypted.

Unfortunately, doing either of those things (returning a blanket 200 and
removing response content) can't be done with "regular" Nginx, so we're
pulling in the OpenResty Lua module to help with that.