Add FORCE_HEALTHCHECK_SUCCESS
This lets us configure Nginx to always return success on a healthcheck.

The underlying goal here is to ensure that if we only have one Nginx
instance running (e.g. v1 or app scaled to 1 instance), we'll never
re-route traffic to Brickwall unless Nginx itself is down.

This has two benefits:

- First, we avoid having to wait for DNS propagation to re-route traffic
  to Brickwall. This (arguably) isn't a major concern in a production
  set up (at this point there are bigger worries), but it definitly
  provides better UX in development (so that we can serve the error page
  as soon as the app crashes).
- Second, Nginx will always do a better job than Brickwall to serve the
  error page: It knows exactly which app was started, which cert should
  be used, etc., whereas Brickwall has to do a lot of guesswork based on
  a SNI Host / Host header, which isn't guaranteed to be successful.

This commit also adds a bit of debugging to our test runner. The tests
for this image are a little flakey, and it's good to have some debugging
input in place to troubleshoot them as they occur.